+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix #184: Make env headers comments for vagrant service-manager env @bexelbie
 - Fix #135: Refactor command.rb to make commands easier to add/maintain @budhrg
 - Adds @budhrg as co-maintainer for the plugin @navidshaikh
+- Fix #191: 'vagrant service-manager restart' not handled correctly @budhrg
 
 ## v1.0.1 Apr 12, 2016
 - Updated SPEC (v1.0.0) for url, date and format @budhrg